# fix(color-contrast): greatly improve color-contrast-matches speed. add aria/get-accessible-ref