--- a/lib/commons/aria/get-accessible-refs.js
+++ b/lib/commons/aria/get-accessible-refs.js
@@ -3,11 +3,11 @@ import cache from '../../core/base/cache';
 import cacheIdRefs from './cache-idrefs';
 
 /**
- * Check that a DOM node is a reference in the accessibility tree
+ * Return all DOM nodes that use the nodes ID in the accessibility tree.
  * @param {Element} node
- * @returns {Boolean}
+ * @returns {Element[]}
  */
-function isAccessibleRef(node) {
+function getAccessibleRefs(node) {
 	node = node.actualNode || node;
 	let root = getRootNode(node);
 	root = root.documentElement || root; // account for shadow roots
@@ -19,7 +19,7 @@ function isAccessibleRef(node) {
 		cacheIdRefs(root);
 	}
 
-	return !!cache.get('idRefs')[id];
+	return cache.get('idRefs')[id] || [];
 }
 
-export default isAccessibleRef;
+export default getAccessibleRefs;