--- a/lib/commons/aria/index.js
+++ b/lib/commons/aria/index.js
@@ -6,6 +6,7 @@
 export { default as allowedAttr } from './allowed-attr';
 export { default as arialabelText } from './arialabel-text';
 export { default as arialabelledbyText } from './arialabelledby-text';
+export { default as getAccessibleRefs } from './get-accessible-refs';
 export { default as getElementUnallowedRoles } from './get-element-unallowed-roles';
 export { default as getExplicitRole } from './get-explicit-role';
 export { default as getOwnedVirtual } from './get-owned-virtual';