--- a/lib/commons/aria/is-accessible-ref.js
+++ b/lib/commons/aria/is-accessible-ref.js
@@ -1,6 +1,4 @@
-import getRootNode from '../dom/get-root-node';
-import cache from '../../core/base/cache';
-import cacheIdRefs from './cache-idrefs';
+import getAccessibleRefs from './get-accessible-refs';
 
 /**
  * Check that a DOM node is a reference in the accessibility tree
@@ -8,18 +6,7 @@ import cacheIdRefs from './cache-idrefs';
  * @returns {Boolean}
  */
 function isAccessibleRef(node) {
-	node = node.actualNode || node;
-	let root = getRootNode(node);
-	root = root.documentElement || root; // account for shadow roots
-	const id = node.id;
-
-	// because axe.commons is not available in axe.utils, we can't do
-	// this caching when we build up the virtual tree
-	if (!cache.get('idRefs')) {
-		cacheIdRefs(root);
-	}
-
-	return !!cache.get('idRefs')[id];
+	return !!getAccessibleRefs(node).length;
 }
 
 export default isAccessibleRef;