--- a/lib/rules/color-contrast-matches.js
+++ b/lib/rules/color-contrast-matches.js
@@ -1,12 +1,9 @@
 /* global document */
+import { getAccessibleRefs } from '../commons/aria';
 import { findUpVirtual, visuallyOverlaps, getRootNode } from '../commons/dom';
 import { visibleVirtual, removeUnicode, sanitize } from '../commons/text';
 import { isDisabled } from '../commons/forms';
-import {
-	getNodeFromTree,
-	querySelectorAll,
-	escapeSelector
-} from '../core/utils';
+import { getNodeFromTree, querySelectorAll } from '../core/utils';
 
 function colorContrastMatches(node, virtualNode) {
 	const { nodeName, type: inputType } = virtualNode.props;
@@ -96,11 +93,7 @@ function colorContrastMatches(node, virtualNode) {
 	while (ancestorNode) {
 		// Find any ancestor (including itself) that is used with aria-labelledby
 		if (ancestorNode.props.id) {
-			const doc = getRootNode(node);
-			const escapedId = escapeSelector(ancestorNode.props.id);
-			const controls = Array.from(
-				doc.querySelectorAll(`[aria-labelledby~="${escapedId}"]`)
-			);
+			const controls = getAccessibleRefs(ancestorNode);
 			const virtualControls = controls.map(control => getNodeFromTree(control));
 
 			ariaLabelledbyControls.push(...virtualControls);