--- a/lib/rules/color-contrast-matches.js
+++ b/lib/rules/color-contrast-matches.js
@@ -3,7 +3,7 @@ import { getAccessibleRefs } from '../commons/aria';
 import { findUpVirtual, visuallyOverlaps, getRootNode } from '../commons/dom';
 import { visibleVirtual, removeUnicode, sanitize } from '../commons/text';
 import { isDisabled } from '../commons/forms';
-import { getNodeFromTree, querySelectorAll } from '../core/utils';
+import { getNodeFromTree, querySelectorAll, tokenList } from '../core/utils';
 
 function colorContrastMatches(node, virtualNode) {
 	const { nodeName, type: inputType } = virtualNode.props;
@@ -93,8 +93,13 @@ function colorContrastMatches(node, virtualNode) {
 	while (ancestorNode) {
 		// Find any ancestor (including itself) that is used with aria-labelledby
 		if (ancestorNode.props.id) {
-			const controls = getAccessibleRefs(ancestorNode);
-			const virtualControls = controls.map(control => getNodeFromTree(control));
+			const virtualControls = getAccessibleRefs(ancestorNode)
+				.filter(control => {
+					return tokenList(control.getAttribute('aria-labelledby')).includes(
+						ancestorNode.props.id
+					);
+				})
+				.map(control => getNodeFromTree(control));
 
 			ariaLabelledbyControls.push(...virtualControls);
 		}