--- a/lib/rules/color-contrast-matches.js
+++ b/lib/rules/color-contrast-matches.js
@@ -95,9 +95,9 @@ function colorContrastMatches(node, virtualNode) {
 		if (ancestorNode.props.id) {
 			const virtualControls = getAccessibleRefs(ancestorNode)
 				.filter(control => {
-					return tokenList(control.getAttribute('aria-labelledby')).includes(
-						ancestorNode.props.id
-					);
+					return tokenList(
+						control.getAttribute('aria-labelledby') || ''
+					).includes(ancestorNode.props.id);
 				})
 				.map(control => getNodeFromTree(control));