The color-contrast-matches function would try to query the entire DOM for every node with an `id` attribute, and all children of that node would also do the same check. This caused the matcher to take ~40s on a relatively small page with lots of IDs. Since we were already caching id references in `aria/is-accessible-ref`, I was able to leverage that same code to create a new function that returns an array of all id ref DOM nodes. That way we could just reference that cache in color-contrast-matches and reduce the time to about 1 second.

Here are the perf results in running only the color-contrast rule on https://www.chase.com/content/dam/chaseonline/en/demos/cbo/launch.html.

| Time Before (ms) | Time After (ms) |
| - | - |
| 42947.73 | 1271.18 |

**NOTE:** We are in a code freeze so cannot merge this until after 4.1 is released on Monday.

Closes issue: #2479 

## Reviewer checks

**Required fields, to be filled out by PR reviewer(s)**

- [x] Follows the commit message policy, appropriate for next version
- [x] Code is reviewed for security